Create mystopeast
A terrible piece of code which I have done spring boarding off Dave Beasley's PyCon 2013 workshop.  It monitors the buses coming to the stop I use most as a beginning point for a trip.  Anyone though can easily take this code and change it to their own route and stop.